LayoutTests/svg/animations/animate-color-fill-currentColor-expected.txt

CONSOLE MESSAGE: line 2: ReferenceError: Can't find variable: createSVGTestCase
SVG 1.1 dynamic animation tests

Tests animation on 'currentColor'.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".
